backend/requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
packaging==23.1
psycopg2==2.9.7
pycodestyle==2.11.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
uWSGI==2.0.22
